requirements.txt
Flask>=2.2
flask-login>=0.6
flask-sqlalchemy>=3.0
requests
sqlalchemy
alembic
python-dotenv
google-generativeai>=0.7.0
huggingface-hub>=0.18.0
